Refactor tokenization and ranges more generically
Enable string tokenization and parsing of integer ranges using more
generic free functions in egs_input. While this was first implemented in
the setLabels method, it is more generally useful to expand hyphenated
ranges of integers, for example in region lists across EGSnrc. Also add
a region number validation method is in the base geometry class.
Finally, change the label type to EGS_Label, for consistency.