Fix isotropic collimation by angle check
Fix egs_isotropic_source collimation using theta or phi ranges so that
the code exits if the user tries to do a range over the branch point.
There is now a simple requirement that max_phi > min_phi and max_theta >
min_theta to avoid the user making this mistake.